+++ src/aks-preview/azext_aks_preview/_client_factory.py
@@ -37,6 +37,8 @@ def cf_managed_clusters(cli_ctx, *_):
 def cf_agent_pools(cli_ctx, *_):
     return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS_PREVIEW).agent_pools
 
+def cf_snapshots(cli_ctx, *_):
+    return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS_PREVIEW).snapshots
 
 def cf_maintenance_configurations(cli_ctx, *_):
     return get_mgmt_service_client(cli_ctx, CUSTOM_MGMT_AKS_PREVIEW).maintenance_configurations