+++ src/aks-preview/azext_aks_preview/_validators.py
@@ -504,3 +504,8 @@ def validate_assign_kubelet_identity(namespace):
         from msrestazure.tools import is_valid_resource_id
         if not is_valid_resource_id(namespace.assign_kubelet_identity):
             raise CLIError("--assign-kubelet-identity is not a valid Azure resource ID.")
+
+def validate_source_nodepool_id(namespace):
+    from msrestazure.tools import is_valid_resource_id
+    if not is_valid_resource_id(namespace.source_nodepool_id):
+        raise CLIError("--source-nodepool-id is not a valid Azure resource ID.")